refactor: remove Pokémon commands
