+++ src/.env
@@ -50,7 +50,6 @@ SAELEM_ENABLED=true
 SAELEM_URL='http://localhost:8284'
 
 OUTFLUX_URL='http://localhost:8286'
-GRAPHQL_POKEMON_URL='https://graphqlpokemon.favware.tech'
 GRAPHQL_STARWARS_URL='https://swapi.skyra.pw'
 HASTEBIN_POST_URL='https://hastebin.skyra.pw'
 HASTEBIN_GET_URL='https://hastebin.skyra.pw'
