--- src/languages/es-ES/commands/pokemon.json
@@ -1,298 +1,0 @@
-{
-	"dragoniteReminder": "",
-	"abilityDescription": "",
-	"abilityExtended": {
-		"usages": [
-			""
-		],
-		"extendedHelp": "",
-		"explainedUsage": [
-			[
-				"",
-				""
-			]
-		],
-		"examples": [
-			"",
-			""
-		]
-	},
-	"abilityEmbedTitles": {
-		"authorTitle": "",
-		"fieldEffectTitle": ""
-	},
-	"abilityQueryFail": "",
-	"flavorsDescription": "",
-	"flavorsExtended": {
-		"usages": [
-			"",
-			"",
-			"",
-			""
-		],
-		"extendedHelp": "",
-		"explainedUsage": [
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			]
-		],
-		"examples": [
-			"",
-			"",
-			"",
-			"",
-			""
-		]
-	},
-	"flavorsQueryFail": "",
-	"flavorNoFlavors": "",
-	"itemDescription": "",
-	"itemExtended": {
-		"usages": [
-			""
-		],
-		"extendedHelp": "",
-		"explainedUsage": [
-			[
-				"",
-				""
-			]
-		],
-		"examples": [
-			"",
-			""
-		]
-	},
-	"itemEmbedData": {
-		"ITEM": "",
-		"generationIntroduced": "",
-		"availableInGeneration8Title": "",
-		"availableInGeneration8Data": ""
-	},
-	"itemQueryFail": "",
-	"learnDescription": "",
-	"learnExtended": {
-		"usages": [
-			"",
-			"",
-			"",
-			"",
-			"",
-			""
-		],
-		"extendedHelp": "",
-		"explainedUsage": [
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			]
-		],
-		"examples": [
-			"",
-			"",
-			"",
-			"",
-			"",
-			""
-		]
-	},
-	"learnMethodTypes": {
-		"levelUpMoves": "",
-		"eventMoves": "",
-		"tutorMoves": "",
-		"eggMoves": "",
-		"virtualTransferMoves": "",
-		"tmMoves": "",
-		"dreamworldMoves": ""
-	},
-	"learnInvalidGeneration": "",
-	"learnMethod": "",
-	"learnQueryFailed": "",
-	"learnCannotLearn": "",
-	"learnTitle": "",
-	"moveDescription": "",
-	"moveExtended": {
-		"usages": [
-			""
-		],
-		"extendedHelp": "",
-		"explainedUsage": [
-			[
-				"",
-				""
-			]
-		],
-		"examples": [
-			"",
-			"",
-			""
-		],
-		"reminder": ""
-	},
-	"moveEmbedData": {
-		"move": "",
-		"types": "",
-		"basePower": "",
-		"pp": "",
-		"category": "",
-		"accuracy": "",
-		"priority": "",
-		"target": "",
-		"contestCondition": "",
-		"zCrystal": "",
-		"gmaxPokemon": "",
-		"availableInGeneration8Title": "",
-		"availableInGeneration8Data": "",
-		"none": "",
-		"maxMovePower": "",
-		"zMovePower": "",
-		"fieldMoveEffectTitle": ""
-	},
-	"moveQueryFail": "",
-	"pokedexDescription": "",
-	"pokedexExtended": {
-		"usages": [
-			"",
-			"",
-			"",
-			""
-		],
-		"extendedHelp": "",
-		"explainedUsage": [
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			]
-		],
-		"examples": [
-			"",
-			"",
-			"",
-			"",
-			""
-		],
-		"reminder": ""
-	},
-	"pokedexEmbedData": {
-		"types": "",
-		"abilities": "",
-		"genderRatio": "",
-		"smogonTier": "",
-		"uknownSmogonTier": "",
-		"height": "",
-		"weight": "",
-		"eggGroups": "",
-		"evolutionaryLine": "",
-		"baseStats": "",
-		"baseStatsTotal": "",
-		"evYields": "",
-		"flavourText": "",
-		"otherFormesTitle": "",
-		"cosmeticFormesTitle": "",
-		"otherFormesList": "",
-		"cosmeticFormesList": "",
-		"levellingRate": "",
-		"isEggObtainable": "",
-		"minimumHatchingTime": "",
-		"maximumHatchTime": ""
-	},
-	"pokedexQueryFail": "",
-	"spriteDescription": "",
-	"spriteExtended": {
-		"usages": [
-			"",
-			"",
-			"",
-			""
-		],
-		"extendedHelp": "",
-		"explainedUsage": [
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			],
-			[
-				"",
-				""
-			]
-		],
-		"examples": [
-			"",
-			"",
-			"",
-			"",
-			""
-		]
-	},
-	"typeDescription": "",
-	"typeExtended": {
-		"usages": [
-			"",
-			""
-		],
-		"extendedHelp": "",
-		"explainedUsage": [
-			[
-				"",
-				""
-			]
-		],
-		"examples": [
-			"",
-			""
-		]
-	},
-	"typeEmbedData": {
-		"offensive": "",
-		"defensive": "",
-		"superEffectiveAgainst": "",
-		"dealsNormalDamageTo": "",
-		"doesNotAffect": "",
-		"notVeryEffectiveAgainst": "",
-		"vulnerableTo": "",
-		"takesNormalDamageFrom": "",
-		"resists": "",
-		"notAffectedBy": "",
-		"typeEffectivenessFor": ""
-	},
-	"typeTooManyTypes": "",
-	"typeNotAType": "",
-	"typeQueryFail": ""
-}
